BIP script restoration: increase stack depth limit to 32768.
OP_TX can require this for real txs: 1000 can start to become limiting
if we want to put values from every output or input on the stack.

This also brings the issue of OP_ROLL to the fore: as mentioned in BIP-342,
it can be time-consuming, and this gets worse as we increase stack size.

After consideration and initial benchmarking, we cost this as if it
were doing a copy of 24 bytes per stack operation moved, which is how
it works in the current bitcoind C++ code.  A more efficient
implementation is possible, but large OP_ROLLs are not something which
are done in practice, so minimizing the cost (or unduly optimizing it)
is unnecessary.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>